pyproject.toml
[build-system]
requires = ["poetry>=1.0.5"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "stjudecloud-merge-counts"
version = "1.0.0"
description = "Utility for merging RNA-seq expression counts files from St. Jude Cloud."
license = "MIT"
authors = [
    "Clay McLeod <Clay.McLeod@STJUDE.org>",
    "Andrew Frantz <Andrew.Frantz@STJUDE.org"
]
readme = "README.md"
repository = "https://github.com/stjudecloud/merge-counts"
keywords = [
    "bioinformatics",
    "genomics",
    "htseq",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

packages = [
    { include = "mergecounts" }
]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.1.0"
tables = "^3.6.1"
tqdm = "^4.48.2"
logzero = "^1.5.0"
dxpy = "^0.298.1"
requests = "<2.24.0"
"hurry.filesize" = "^0.9"

[tool.poetry.dev-dependencies]
mypy = "^0.782"
pylint = "^2.5.3"
black = "^19.10b0"
python-semantic-release = "^7.2.2"

[tool.poetry.scripts]
stjudecloud-merge-counts = 'mergecounts.__main__:run'

[tool.semantic_release]
version_variable = "pyproject.toml:version"
version_source = "commit"
commit_subject = "chore: bumping version to v{version}"
commit_message = ""
commit_author = "St. Jude Cloud <support@stjude.cloud>"
build_command = "poetry build"
upload_to_release = "false"
# check_build_status = true

[tool.black]
target-version = ['py36', 'py37', 'py38']
